fix: #432 Error when logout in the command-line tool

+++ mods/ctl/src/commands/auth/logout.ts
@@ -13,14 +13,15 @@ export default class extends Command {
 
   async run() {
     CliUx.ux.action.start(`Login out`);
-
-    try {
-      fs.rmSync(BASE_DIR, {recursive: true});
-      await CliUx.ux.wait(1000);
-      CliUx.ux.action.stop("Done");
-    } catch (e) {
-      CliUx.ux.action.stop();
-      throw new CLIError(e.message);
+    if (fs.existsSync(BASE_DIR)) {
+      try {
+        fs.rmSync(BASE_DIR, {recursive: true});
+        await CliUx.ux.wait(1000);
+      } catch (e) {
+        CliUx.ux.action.stop();
+        throw new CLIError(e.message);
+      }
     }
+    CliUx.ux.action.stop("Done");
   }
 }
